add OAuthBase class for 3rd party OAuth apps
