+++ dash_auth/plotly_auth.py
@@ -1,24 +1,19 @@
 from __future__ import absolute_import
-import datetime
 import flask
-from flask_seasurf import SeaSurf
 import json
-import os
 from six import iteritems
-from .auth import Auth
+
+from .oauth import OAuthBase
 
 from . import api_requests
 
-AUTH_COOKIE_NAME = 'plotly_auth'
 
+class PlotlyAuth(OAuthBase):
+    AUTH_COOKIE_NAME = 'plotly_auth'
+    TOKEN_COOKIE_NAME = 'plotly_oauth_token'
 
-class PlotlyAuth(Auth):
     def __init__(self, app, app_name, sharing, app_url):
-        Auth.__init__(self, app)
-
-        self.config = {
-            'permissions_cache_expiry': 5 * 60
-        }
+        super(PlotlyAuth, self).__init__(app, app_url)
 
         self._fid = create_or_overwrite_dash_app(
             app_name, sharing, app_url
@@ -26,102 +21,7 @@ class PlotlyAuth(Auth):
         self._oauth_client_id = create_or_overwrite_oauth_app(
             app_url, app_name
         )['client_id']
-
-        self._app = app
-        self._app_url = app_url
         self._sharing = sharing
-        self._access_codes = self.create_access_codes()
-
-        app.server.add_url_rule(
-            '{}_dash-login'.format(app.config['routes_pathname_prefix']),
-            view_func=self.login_api,
-            methods=['post']
-        )
-
-        app.server.add_url_rule(
-            '{}_oauth-redirect'.format(app.config['routes_pathname_prefix']),
-            view_func=self.serve_oauth_redirect,
-            methods=['get']
-        )
-
-        app.server.add_url_rule(
-            '{}_is-authorized'.format(app.config['routes_pathname_prefix']),
-            view_func=self.check_if_authorized,
-            methods=['get']
-        )
-        _current_path = os.path.dirname(os.path.abspath(__file__))
-
-        # TODO - Dist files
-        with open(os.path.join(_current_path, 'oauth-redirect.js'), 'r') as f:
-            self.oauth_redirect_bundle = f.read()
-
-        with open(os.path.join(_current_path, 'login.js'), 'r') as f:
-            self.login_bundle = f.read()
-
-    def create_access_codes(self):
-        token = SeaSurf()._generate_token()
-        new_access_codes = {
-            'access_granted': token,
-            'expiration': (
-                datetime.datetime.now() + datetime.timedelta(
-                    seconds=self.config['permissions_cache_expiry']
-                )
-            )
-        }
-        self._access_codes = new_access_codes
-        return self._access_codes
-
-    def is_authorized(self):
-        if 'plotly_oauth_token' not in flask.request.cookies:
-            return False
-
-        oauth_token = flask.request.cookies['plotly_oauth_token']
-
-        if (datetime.datetime.now() > self._access_codes['expiration']):
-            self.create_access_codes()
-
-        if AUTH_COOKIE_NAME not in flask.request.cookies:
-            return check_view_access(oauth_token, self._fid)
-
-        access_cookie = flask.request.cookies[AUTH_COOKIE_NAME]
-
-        # If there access was previously declined,
-        # check access again in case it has changed
-        if access_cookie != self._access_codes['access_granted']:
-            return check_view_access(oauth_token, self._fid)
-
-        return True
-
-    def check_if_authorized(self):
-        if self.is_authorized():
-            return flask.Response(status=200)
-
-        return flask.Response(status=403)
-
-    def auth_wrapper(self, f):
-        def wrap(*args, **kwargs):
-            if not self.is_authorized():
-                return flask.Response(status=403)
-
-            response = f(*args, **kwargs)
-            # TODO - should set secure in this cookie, not exposed in flask
-            # TODO - should set path or domain
-            try:
-                # Python 2
-                if isinstance(response, basestring):  # noqa: F821
-                    response = flask.Response(response)
-            except:
-                # Python 3
-                if isinstance(response, str):
-                    response = flask.Response(response)
-            self.set_cookie(
-                response,
-                name=AUTH_COOKIE_NAME,
-                value=self._access_codes['access_granted'],
-                max_age=(60 * 60 * 24 * 7),  # 1 week
-            )
-            return response
-        return wrap
 
     def html(self, script):
         return ('''
@@ -149,12 +49,6 @@ class PlotlyAuth(Auth):
             script)
         )
 
-    def login_request(self):
-        return self.html(self.login_bundle)
-
-    def serve_oauth_redirect(self):
-        return self.html(self.oauth_redirect_bundle)
-
     def login_api(self):
         oauth_token = flask.request.get_json()['access_token']
         res = api_requests.get(
@@ -181,14 +75,18 @@ class PlotlyAuth(Auth):
 
         return response
 
-    def set_cookie(self, response, name, value, max_age):
-        response.set_cookie(
-            name,
-            value=value,
-            max_age=max_age,
-            secure=True if 'https:' in self._app_url else False,
-            path=self._app.config['routes_pathname_prefix']
+    def check_view_access(self, oauth_token):
+        res = api_requests.get(
+            '/v2/files/{}'.format(self._fid),
+            headers={'Authorization': 'Bearer {}'.format(oauth_token)}
         )
+        if res.status_code == 200:
+            return True
+        elif res.status_code == 404:
+            return False
+        else:
+            # TODO - Dash exception
+            raise Exception('Failed request to plotly')
 
 
 def create_or_overwrite_dash_app(filename, sharing, app_url):
@@ -287,17 +185,3 @@ def create_or_overwrite_oauth_app(app_url, name):
         raise e
 
     return res.json()
-
-
-def check_view_access(oauth_token, fid):
-    res = api_requests.get(
-        '/v2/files/{}'.format(fid),
-        headers={'Authorization': 'Bearer {}'.format(oauth_token)}
-    )
-    if res.status_code == 200:
-        return True
-    elif res.status_code == 404:
-        return False
-    else:
-        # TODO - Dash exception
-        raise Exception('Failed request to plotly')
